+++ Code/GeneSimulation_py/cppSocialChoiceGenes/Chromosome.h
@@ -12,18 +12,18 @@
 class Chromosome {
 public:
     // Constructor
-    Chromosome(const std::vector<float>& chromosome);
+    Chromosome(const std::vector<double>& chromosome);
 
     // Setters and Getters
-    void setFitness(float fitness);
-    void addFitness(float newFitness);
+    void setFitness(double fitness);
+    void addFitness(double newFitness);
     int getFitness() const;
 
-    std::vector<float> getChromosome() const;
+    std::vector<double> getChromosome() const;
     void resetFitness();
 
     // Operator Overloading
-    float operator[](size_t index) const;  // Access chromosome element
+    double operator[](size_t index) const;  // Access chromosome element
     bool operator==(const Chromosome& other) const;  // Compare chromosomes
 
     // Hashing for the Chromosome (using std::hash)
@@ -32,8 +32,8 @@ public:
     };
 
 private:
-    std::vector<float> chromosome;  // Stores the chromosome
-    float fitness;  // Stores the fitness value
+    std::vector<double> chromosome;  // Stores the chromosome
+    double fitness;  // Stores the fitness value
 };
